scripts/requirements.txt
pandas
polars-lts-cpu